RaspberryScripts/requirements.txt
astroid==2.4.2
azure-iot-device==2.4.0
bluepy==1.3.0
btlewrap==0.0.10
certifi==2020.11.8
chardet==3.0.4
deprecation==2.1.0
idna==2.10
isort==5.6.4
janus==0.4.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
miflora==0.7.1
packaging==20.4
paho-mqtt==1.5.1
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
requests==2.25.0
requests-unixsocket==0.2.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.11
wrapt==1.12.1